Move content blocks directory
